[refactor] 디렉토리 구조 수정
- 폴더명 복수형으로 바꾸고 import
