+++ backend/controllers/auth-controller.js
@@ -1,25 +1,27 @@
 const passport = require('passport');
 
 exports.loginGitHub = (req, res, next) => {
-     passport.authenticate('github', function(err, user, info) {
-         if (err) { return next(err); }
-         
-         // redirect 주소 수정 필요
-         if (!user) { return res.redirect('http://localhost:8080'); }
-         
-         /**
-          * 토큰 생성 부분 넣으면 될 듯
-          * req.logIn(user, function(err) {
-          *     if (err) { return next(err); }
-          *     return res.redirect('/users/' + user.username);
-          * }); 
-          */
+  passport.authenticate('github', function (err, user, info) {
+    if (err) {
+      return next(err);
+    }
 
-         res.cookie('user', user);
+    // redirect 주소 수정 필요
+    if (!user) {
+      return res.redirect('http://localhost:8080');
+    }
 
-         // redirect 주소 수정 필요
-         res.redirect('http://localhost:8080');
+    /**
+     * 토큰 생성 부분 넣으면 될 듯
+     * req.logIn(user, function(err) {
+     *     if (err) { return next(err); }
+     *     return res.redirect('/users/' + user.username);
+     * });
+     */
 
-    })(req, res, next);
+    res.cookie('user', user);
+
+    // redirect 주소 수정 필요
+    res.redirect('http://localhost:8080');
+  })(req, res, next);
 };
-
