+++ backend/controllers/test-controller.js
@@ -1,7 +1,6 @@
 const service = require('../services/test-service');
 
-exports.addUser = async(req, res, next) => {
-     const insertId = await service.userService(); 
-     res.json(insertId); 
+exports.addUser = async (req, res, next) => {
+  const insertId = await service.userService();
+  res.json(insertId);
 };
-
