+++ backend/loaders/express-loader.js
@@ -5,10 +5,10 @@ const logger = require('morgan');
 const cors = require('cors');
 
 module.exports = app => {
-    app.use(cors());
-    app.use(logger('dev'));
-    app.use(express.json());
-    app.use(express.urlencoded({ extended: false }));
-    app.use(cookieParser());
-    app.use(express.static(path.join(__dirname, '../public')));
+  app.use(cors());
+  app.use(logger('dev'));
+  app.use(express.json());
+  app.use(express.urlencoded({ extended: false }));
+  app.use(cookieParser());
+  app.use(express.static(path.join(__dirname, '../public')));
 };
