+++ backend/loaders/passport-loader.js
@@ -1,5 +1,5 @@
 const passportConfig = require('../passport');
 
 module.exports = () => {
-    passportConfig(); 
+  passportConfig();
 };
