+++ backend/services/test-service.js
@@ -1,7 +1,6 @@
-const users = require('../model/users');
+const users = require('../models/users');
 
-exports.userService = async() => {
-    const insertId = await users.create('zlrlo', 'urlurlurl', 'password2');
-    return insertId;      
+exports.userService = async () => {
+  const insertId = await users.create('zlrlo', 'urlurlurl', 'password2');
+  return insertId;
 };
-
